Move internals to `source/internal/function.d.ts`, fix docs
- Move `HasExplicitThis`, `LastOverload`, `CollectOverloads`, and
  internal `Parameters` to `source/internal/function.d.ts`
- Deduplicate JSDoc: TS overload enumeration behavior documented once
  in the internal file; `HasExplicitThis` and `CollectOverloads` refer
  to it instead of repeating
- Fix stale "in this file" reference in `FunctionOverloads` JSDoc
- Add "Known limitations are the same as FunctionOverloads" to
  `OverloadsToTuple` JSDoc
